feat(Implementar Request) :sparkles: Adicionar classes PersonDTO e SplitInvoiceRequest
Foi introduzido um novo objeto de transferência de dados (DTO) para a classe "Pessoa", bem como uma nova classe de requisição para gerenciar a divisão de faturas. A classe enum "DiscountType" foi atualizada e agora inclui métodos para realizar cálculos de desconto. Além disso, um novo atributo foi adicionado ao modelo "Person" com o objetivo de marcar um indivíduo como comprador.
